working shiny speech app

+++ inst/shiny/capture_speech/server.R
@@ -1,15 +1,14 @@
 library(shiny)
 library(tuneR)
 library(googleLanguageR)
+library(shinyjs)
 
-# Define server logic required to draw a histogram
 function(input, output, session){
 
   input_audio <- reactive({
     req(input$audio)
     a <- input$audio
 
-    saveRDS(a, "raw_audio.rds")
     if(length(a) > 0){
       return(a)
     } else {
@@ -28,6 +27,7 @@ function(input, output, session){
     a1 <- a[1:audio_split]
     a2 <- a[(audio_split+1):length(a)]
 
+    # construct wav object that the API likes
     Wobj <- Wave(a1, a2, samp.rate = 44100, bit = 16)
     Wobj <- normalize(Wobj, unit = "16", pcm = TRUE)
     Wobj <- mono(Wobj)
@@ -35,9 +35,7 @@ function(input, output, session){
     wav_name <- paste0("audio",gsub("[^0-9]","",Sys.time()),".wav")
 
     writeWave(Wobj, wav_name, extensible = FALSE)
-    message("saved wav to ", wav_name)
 
-    # "audio20171123132816.wav"
     wav_name
 
 
@@ -46,21 +44,36 @@ function(input, output, session){
 
   output$result_text <- renderText({
     req(wav_name())
+    req(input$language)
+
+    if(input$language == ""){
+      stop("Must enter a languageCode - default en-US")
+    }
 
     wav_name <- wav_name()
 
+    if(!file.exists(wav_name)){
+      return(NULL)
+    }
+
     message("Calling API")
-    me <- googleLanguageR::gl_speech(wav_name, sampleRateHertz = 44100L)
-str(me)
+    shinyjs::show(id = "api", anim = TRUE, animType = "fade", time = 1)
+
+    # make API call
+    me <- gl_speech(wav_name,
+                    sampleRateHertz = 44100L,
+                    languageCode = input$language)
+
+    ## remove old file
+    unlink(wav_name)
 
     message("API returned: ", me$transcript$transcript)
-    unlink(wav_name)
+    shinyjs::hide(id = "api", anim = TRUE, animType = "fade", time = 1)
 
     as.character(me$transcript$transcript)
 
   })
 
-  output$the_time <- renderText({as.character(Sys.time())})
 
 
 }
